ForecastAccuracy.jl

Implementation of forecast accuracy measures.

Installation

To install the package you can use Pkg.add it as follows:

pkg> add ForecastAccuracy

Contributing

Contributions to this package are more than welcome, if you find a bug or have any suggestions for the documentation please post it on the github issue tracker.

When contributing please note that the package follows the JuMP style guide.

Accuracy measures

ForecastAccuracy.me — Function

me(real::Vector{T}, forecast::Vector{T}) where T

Mean Error.

ForecastAccuracy.mde — Function

mde(real::Vector{T}, forecast::Vector{T}) where T

Median Error.

ForecastAccuracy.mae — Function

mae(real::Vector{T}, forecast::Vector{T}) where T

Mean Absolute Error.

ForecastAccuracy.gmae — Function

gmae(real::Vector{T}, forecast::Vector{T}) where T

Geometric Mean Absolute Error.

ForecastAccuracy.mdae — Function

mdae(real::Vector{T}, forecast::Vector{T}) where T

Median Absolute Error.

ForecastAccuracy.mse — Function

mse(real::Vector{T}, forecast::Vector{T}) where T

Mean Square Error.

ForecastAccuracy.mdse — Function

mdse(real::Vector{T}, forecast::Vector{T}) where T

Median Square Error.

ForecastAccuracy.rmse — Function

rmse(real::Vector{T}, forecast::Vector{T}) where T

Root Mean Squared Error.

ForecastAccuracy.mpe — Function

mpe(real::Vector{T}, forecast::Vector{T}) where T

Mean Percentual Error.

ForecastAccuracy.mdpe — Function

mdpe(real::Vector{T}, forecast::Vector{T}) where T

Median Percentual Error.

ForecastAccuracy.mape — Function

mape(real::Vector{T}, forecast::Vector{T}) where T

Mean Absolute Percentual Error.

ForecastAccuracy.mdape — Function

mdape(real::Vector{T}, forecast::Vector{T}) where T

Median Absolute Percentual Error.

ForecastAccuracy.smape — Function

smape(real::Vector{T}, forecast::Vector{T}) where T

Symmetric Mean Absolute Percentual Error.

ForecastAccuracy.smdape — Function

smdape(real::Vector{T}, forecast::Vector{T}) where T

Symmetric Median Absolute Percentual Error.

ForecastAccuracy.maape — Function

maape(real::Vector{T}, forecast::Vector{T}) where T

Mean Arctangent Absolute Percentage Error. As developed in https://www.sciencedirect.com/science/article/pii/S0169207016000121

ForecastAccuracy.mase — Function

mase(real::Vector{T}, forecast::Vector{T}; steps_ahead::Int = 1) where T

Mean Absolute Scaled Error. As developed in https://robjhyndman.com/papers/mase.pdf

ForecastAccuracy.mdase — Function

mdase(real::Vector{T}, forecast::Vector{T}; steps_ahead::Int = 1) where T

Median Absolute Scaled Error. As developed in https://robjhyndman.com/papers/mase.pdf

ForecastAccuracy.rmsse — Function

rmsse(real::Vector{T}, forecast::Vector{T}; steps_ahead::Int = 1) where T

Root Mean Squared Scaled Error.

ForecastAccuracy.pl — Function

pl(real::Vector{T}, quantile_forecast::Vector{T}, tau::T) where T

Pinball loss function defined in terms of the quantile tau